Added snapshot tests for Doku component
