+++ Tests/Components Tests/Doku/DokuComponentTests.swift
@@ -77,130 +77,6 @@ class DokuComponentTests: XCTestCase {
         XCTAssertEqual(sut.button.title, localizedString(LocalizationKey(key: "adyen_confirmPurchase"), sut.configuration.localizationParameters))
     }
 
-    func testUIConfiguration() {
-        var style = FormComponentStyle()
-
-        /// Footer
-        style.mainButtonItem.button.title.color = .white
-        style.mainButtonItem.button.title.backgroundColor = .red
-        style.mainButtonItem.button.title.textAlignment = .center
-        style.mainButtonItem.button.title.font = .systemFont(ofSize: 22)
-        style.mainButtonItem.button.backgroundColor = .red
-        style.mainButtonItem.backgroundColor = .brown
-
-        /// background color
-        style.backgroundColor = .yellow
-
-        /// Text field
-        style.textField.text.color = .brown
-        style.textField.text.font = .systemFont(ofSize: 13)
-        style.textField.text.textAlignment = .right
-
-        style.textField.title.backgroundColor = .blue
-        style.textField.title.color = .yellow
-        style.textField.title.font = .systemFont(ofSize: 20)
-        style.textField.title.textAlignment = .center
-        style.textField.backgroundColor = .red
-
-        let config = DokuComponent.Configuration(style: style)
-        let sut = DokuComponent(paymentMethod: paymentMethod,
-                                context: context,
-                                configuration: config)
-
-        UIApplication.shared.keyWindow?.rootViewController = sut.viewController
-        
-        wait(for: .milliseconds(300))
-        
-        /// Test firstName field
-        self.assertTextInputUI(DokuViewIdentifier.firstName,
-                               view: sut.viewController.view,
-                               style: style.textField,
-                               isFirstField: true)
-
-        /// Test lastName field
-        self.assertTextInputUI(DokuViewIdentifier.lastName,
-                               view: sut.viewController.view,
-                               style: style.textField,
-                               isFirstField: false)
-
-        /// Test email field
-        self.assertTextInputUI(DokuViewIdentifier.email,
-                               view: sut.viewController.view,
-                               style: style.textField,
-                               isFirstField: false)
-
-        /// Test submit button
-        let payButtonItemViewButton: UIControl? = sut.viewController.view.findView(with: DokuViewIdentifier.payButton)
-        let payButtonItemViewButtonTitle: UILabel? = sut.viewController.view.findView(with: "AdyenComponents.DokuComponent.payButtonItem.button.titleLabel")
-
-        XCTAssertEqual(payButtonItemViewButton?.backgroundColor, .red)
-        XCTAssertEqual(payButtonItemViewButtonTitle?.backgroundColor, .red)
-        XCTAssertEqual(payButtonItemViewButtonTitle?.textAlignment, .center)
-        XCTAssertEqual(payButtonItemViewButtonTitle?.textColor, .white)
-        XCTAssertEqual(payButtonItemViewButtonTitle?.font, .systemFont(ofSize: 22))
-    }
-
-    private func assertTextInputUI(_ identifier: String,
-                                   view: UIView,
-                                   style: FormTextItemStyle,
-                                   isFirstField: Bool) {
-
-        let textView: FormTextInputItemView? = view.findView(with: identifier)
-        let textViewTitleLabel: UILabel? = view.findView(with: "\(identifier).titleLabel")
-        let textViewTextField: UITextField? = view.findView(with: "\(identifier).textField")
-
-        XCTAssertEqual(textView?.backgroundColor, .red)
-        XCTAssertEqual(textViewTitleLabel?.textColor, isFirstField ? view.tintColor : style.title.color)
-        XCTAssertEqual(textViewTitleLabel?.backgroundColor, style.title.backgroundColor)
-        XCTAssertEqual(textViewTitleLabel?.textAlignment, style.title.textAlignment)
-        XCTAssertEqual(textViewTitleLabel?.font, style.title.font)
-        XCTAssertEqual(textViewTextField?.backgroundColor, style.backgroundColor)
-        XCTAssertEqual(textViewTextField?.textAlignment, style.text.textAlignment)
-        XCTAssertEqual(textViewTextField?.textColor, style.text.color)
-        XCTAssertEqual(textViewTextField?.font, style.text.font)
-    }
-
-    func testSubmitForm() throws {
-        let sut = DokuComponent(paymentMethod: paymentMethod,
-                                context: context,
-                                configuration: DokuComponent.Configuration())
-        let delegate = PaymentComponentDelegateMock()
-        sut.delegate = delegate
-
-        let delegateExpectation = expectation(description: "PaymentComponentDelegate must be called when submit button is clicked.")
-        delegate.onDidSubmit = { data, component in
-            XCTAssertTrue(component === sut)
-            XCTAssertTrue(data.paymentMethod is DokuDetails)
-            let data = data.paymentMethod as! DokuDetails
-            XCTAssertEqual(data.firstName, "Mohamed")
-            XCTAssertEqual(data.lastName, "Smith")
-            XCTAssertEqual(data.emailAddress, "mohamed.smith@domain.com")
-
-            sut.stopLoadingIfNeeded()
-            delegateExpectation.fulfill()
-            XCTAssertEqual(sut.viewController.view.isUserInteractionEnabled, true)
-            XCTAssertEqual(sut.button.showsActivityIndicator, false)
-        }
-
-        UIApplication.shared.keyWindow?.rootViewController = sut.viewController
-        
-        wait(for: .milliseconds(300))
-        
-        let submitButton: UIControl? = sut.viewController.view.findView(with: DokuViewIdentifier.payButton)
-
-        let firstNameView: FormTextInputItemView! = sut.viewController.view.findView(with: DokuViewIdentifier.firstName)
-        self.populate(textItemView: firstNameView, with: "Mohamed")
-
-        let lastNameView: FormTextInputItemView! = sut.viewController.view.findView(with: DokuViewIdentifier.lastName)
-        self.populate(textItemView: lastNameView, with: "Smith")
-
-        let emailView: FormTextInputItemView! = sut.viewController.view.findView(with: DokuViewIdentifier.email)
-        self.populate(textItemView: emailView, with: "mohamed.smith@domain.com")
-
-        submitButton?.sendActions(for: .touchUpInside)
-        waitForExpectations(timeout: 10, handler: nil)
-    }
-
     func testBigTitle() {
         let sut = DokuComponent(paymentMethod: paymentMethod,
                                 context: context,
